net: Indicate whether a socket is a transparent socket
Modify the printout functions for IPv4-TCP, IPv4-UDP, IPv6-TCP, IPv6-UDP,
such that the state for the socket is printed as state = state | 0x80.
The actual socket state is unmodified. This change is required for the
user space to determine whether a socket is a transparent socket, and
to determine if the socket holder intends to consume packets locally
or to forward them to an external processor.

Signed-off-by: Harout Hedeshian <harouth@codeaurora.org>

Change-Id: I2ca403b4c2c74e7ddcdbda53e4ba43613bae9c42