msm: mdss: support parallel execution of MDSS_EVENT handlers
The existing implementation of mdss_mdp_ctl_intf_event() calls all
subscribed handlers of a given event sequentially. This results in
poor performance in some cases.

For example, some DSI panels have long sleeps (~120ms) in their power-
on sequences. For dual-DSI displays with both panels are subscribing
to the same unblank event, these long sleeps occur one-after-the-other
to the detriment of display-on latency. By modifying
mdss_mdp_ctl_intf_event() to schedule the handlers on worker threads,
commands to both DSI panels can be sent in parallel, with the
associated sleeps occurring in parallel also.

Change-Id: I0b0dcaa6406f3bc4eaffcac4c5f1c505de4de6b1
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>